< HW01 > Update App to ver.2.1. The final version for HW01.
# Description
Limit the scales to specific values. EX: 'kernel_size' ranges from 3 to 15, and
it can only be an integer. Please remind that the 'kernel_size' might be even
number in the scale bar, but even number is invalid. The APP will find that the
value is invalid and shutdown the APP. I haven't fixed it and I won't fix it XD